requirements.txt
fastapi
uvicorn

PyPDF2

langchain
langchain-community
langchain-huggingface
langchain-core
langchain_text_splitters
langchain_google_genai

huggingface-hub
transformers
sentence-transformers

faiss-cpu

pydantic
python-dotenv
